Refactor vigil.py -> main.py
Now that beginnings of a prototype are here, need to split logic before
scaling. Applying single responsiblity principle and trying to move
towards model/view architecture.

Split widgets & such out into new modules.
