+++ generated_ui/splash_ui.py
